renamed subpackage set_up_grasp_models to set_up_models
