+++ docs/set_up_models.rst
@@ -6,5 +6,5 @@ Most of this subpackage is still under construction, only the mechanism conversi
 Convert mechanisms
 -------------------
 
-.. automodule:: set_up_grasp_models.set_up_grasp_models.convert_mechanisms
+.. automodule:: set_up_grasp_models.set_up_models.convert_mechanisms
     :members:
