+++ set_up_grasp_models/set_up_models/convert_mechanisms.py
